sdl13 fix

+++ SDL/SDL13/include/SDL_opengles.h
@@ -113,8 +113,8 @@ void Draw2DStrip(); //AKR
     typedef int GLfixed;
     typedef int GLclampx;
 
-    typedef int GLintptr;
-    typedef int GLsizeiptr;
+    typedef long GLintptr;
+    typedef long GLsizeiptr;
 
 
 /*************************************************************/
